Fix #404: Refactor to use module fido2-model
